requirements.txt
numpy>=2.2.1
pandas>=2.2.3
scikit-learn>=1.5.2 
grpcio>=1.70.0
grpcio-tools>=1.70.0
joblib>=1.4.2
grpcio-reflection>=1.70.0
neuralforecast>=3.0.0
polars>=1.25.2